backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
instaloader==4.10.1
google-generativeai==0.3.2
APScheduler==3.10.4
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.0.3
